+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - `react-dom`,
   - `react-styleguidist`.
 
+- Added iconic style for large Button.
+
 ## [0.27.0] - 2020-10-14
 
 - Couple of small updates in Tooltip and Modal.